Safari-609.1.11.1/LayoutTests/accessibility/ios-simulator/landmark-type-expected.txt

non landmark
text
text
text
footer
article
document
note
This tests that landmarks return their type as part of the AX label.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS main.description is 'AXLabel: main'
PASS main.childAtIndex(0).traits != nonLandmarkText.traits is true
PASS banner.description is 'AXLabel: banner'
PASS banner.childAtIndex(0).traits != nonLandmarkText.traits is true
PASS region.description is 'AXLabel: testregion, region'
PASS region.childAtIndex(0).traits != nonLandmarkText.traits is true
PASS footer.description is 'AXLabel: content information'
PASS footer.childAtIndex(0).traits != nonLandmarkText.traits is true
PASS article.description is 'AXLabel: article'
PASS article.childAtIndex(0).traits != nonLandmarkText.traits is true
PASS note.description is 'AXLabel: note'
PASS note.childAtIndex(0).traits != nonLandmarkText.traits is true
PASS documentContainer.description is 'AXLabel: document'
PASS documentContainer.childAtIndex(0).traits != nonLandmarkText.traits is true
PASS region.childAtIndex(0).traits == main.childAtIndex(0).traits is true
PASS successfullyParsed is true

TEST COMPLETE
